requirements.txt
GPU: NVIDIA GeForce RTX4090 GPU
Pytorch: 1.13.0 cuda 11.7
cudatoolkit: 11.7.1
scikit-learn: 1.0.2
albumentations: 1.2.0



AutoDL
Pytorch
CUDA 12.6

To activate this environment, use
#
#     $ conda activate Medical-Image-Segmentation-Benchmarks-main
#
# To deactivate an active environment, use
#
#     $ conda deactivate


#定位到auto-tmp
cd ~/autodl-tmp/cmu-net

#展示文件
ls

#安装环境
pip install torch torchvision torchaudio

pip install scikit-learn==1.0.2 albumentations==1.2.0

pip install numpy scipy opencv-python PyYAML yacs einops ml-collections timm

